2pc protocol with par cmd (#1014)
This PR accomplishes a proof of the 2pc protocol using a proof rule for
the parallel cmd for summarizing unbounded number of right mover voting
actions. The resulting proof is more succinct that previous attempts
based on pending asyncs. This proof also has the desirable property that
it works for arbitrary number of instance of 2pc potentially running
concurrently.

After this PR lands, this is the only proof of 2PC in the repository.

---------

Co-authored-by: Shaz Qadeer <shaz.qadeer@gmail.com>