Update README.md (#1025)
I am updating the documentation in the `README.md` file to include a
tutorial for Boogie. The link points to the "This is Boogie2" paper. I
am dropping the links to the incomplete documentation and language
reference pages. Even though "This is Boogie2" is slightly out-of-date,
it still appears to be a much more useful document than those pointed to
by the links.