[Civl] Change signature of KeyedLocSet_New (#1029)
This PR changes the signature of KeyedLocSet_New. The new signature and
implementation are as follows:
```
pure procedure {:inline 1} KeyedLocSet_New<K>(ks: Set K) returns ({:linear} {:pool "Loc_New"} l: One Loc, {:linear} keyed_locs: Set (KeyedLoc K))
{
  assume {:add_to_pool "Loc_New", l} true;
  keyed_locs := Set((lambda x: KeyedLoc K :: x->l == l->val && Set_Contains(ks, x->k)));
}
```

The realization is that a linear Loc and a linear collection of
KeyedLoc's can both be returned without any impact on soundness.

The new signature does not change the expressiveness of NPL (the linear
typing system underlying Civl), but it allows us to program without
putting KeyedLoc's into the domain of a linear map. This maps it easy to
erase the code using standard pointer-based memory management.

As a result of this change, the proof of Treiber only requires one
member in the enum LocPiece.

The following changes are done in addition:
- KeyedLoc is changed to TaggedLoc
- a new example two-lists.bpl is added
- a Unit type is added to base.bpl