[Civl] Action preconditions (#1051)
This PR adds a new feature to Civl: preconditions to actions. If the
action is annotated left mover, the precondition is used in proving its
mover checks. Otherwise, the precondition has no effect. These
preconditions are checked in the non-preemptive semantics when the
action is called. It is required that the invariants called in these
preconditions are global invariants. Thus, this PR presents a new way to
achieve inductive sequentialization without dropping down to pending
asyncs. The proof of BroadcastConsensus is revised so that it uses this
new feature and does not use pending asyncs.