[Civl] Refactor linear type checking (#1080)
This is the first PR to start a refactoring of the linear type checking
in Civl. The goal is to handle nesting of linear values properly and to
infer linearity of a variable from its type.

A type is linear if it is either One T or has One T embedded inside it.
Standard (infinite) maps in Boogie are exempt from this definition,
i.e., they are always ordinary. A type is ordinary if it is not linear.
The linearity of a variable is inferred from its type. If a variable's
type is ordinary, it is guaranteed not contain any permissions. Infinite
maps cannot be used to store permissions. But the finite variants Set V
and Map K V introduced by Civl may contain permissions.

A linear variable may be available or unavailable. Ordinary variables
may be considered to be always available or always unavailable, it does
not really matter. The disjointness invariant is guaranteed on all
available linear variables in the program state.

The :linear, :linear_in, and :linear_out attributes are still relevant
but only on globals and parameters. They indicate the expected
availability of the annotated variables in a particular and are needed
to localize type checking. Therefore, it may be useful (later on) to
rename them suitably to emphasize their role only for the purpose of
specifying availability.

TBD:
- Inline constructors if they are not recursive